backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
email-validator
python-dotenv
python-jose
mysqlclient
pymysql